// types/RSSFeedItem.ts
export interface RSSFeedItem {
  creator: string;
  title: string;
  link: string;
  pubDate: string;
  'content:encoded': string;
  'content:encodedSnippet': string;
  'dc:creator': string;
  'content.encoded': string;
  updated: string;
  guid: string;
  categories: string[];
  isoDate: string;
}